Fixed double free bug in the pointcloud compression voxel filter, cleaned up overall structure, added const-correctness, and enabled all warnings/errors.

+++ opportunistic_link/src/image_resizer.cpp
@@ -112,7 +112,7 @@ public:
             intermediate = cv_ptr->image;
         }
         // Resize image
-        if (resized_width_ < image->width && resized_height_ < image->height)
+        if (resized_width_ < (int)(image->width) && resized_height_ < (int)(image->height))
         {
             // If we're resizing smaller, use CV_INTER_AREA interpolation
             cv::resize(intermediate, resized, resized.size(), 0.0, 0.0, CV_INTER_AREA);
